Add PackageLoader to handle package files
The goal is to provide the YAML implementation but pave the way for
a database driven model later by abstracting the loading and parsing.

Two interfaces were chosen to allow for independent revisions. The
file based loader shouldn't change much but the parsers will given the
revisions to the YAML etc.